o1vm/mips: move regtest reg. number of selectors into MIPS directory
It was initially in the pickles subdirectory, which does not make sense as it is
not strictly related to the Pickles flavor.
